Add getBIP32ExtendedKey() command APDU
getBIP32ExtendedKey() returns the BIP32 derived pubkey using a 4-byte counter as path.
At each getBIP32ExtendedKey() call, the counter is incremented, thus a unique pubkey is returned each time.

This pukey can be used as P2PK for locking a token using NUT-11.
The 4-byte counter is sufficient to recover the pubkey from the BIP32 master-key and chaincode.

TODO:
* add optional 4-byte counter in proof storage.
* sign proof using NUT-11 when exporting P2PK token